[fix] moved useTracking's query functionality inside api/tracking.ts
